requirements.txt
cryptography==46.0.5
Flask-Limiter==4.1.1
Flask==3.1.3
google-cloud-compute==1.43.0
google-cloud-secret-manager==2.26.0
google-cloud-tasks==2.16.0
gunicorn==25.1.0
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
